+++ src/Phantonia.Historia.Language/SemanticAnalysis/Binder.cs
@@ -118,7 +118,7 @@ public sealed partial class Binder
     {
         switch (declaration)
         {
-            case SceneSymbolDeclarationNode { Name: string name, Index: int index }:
+            case SceneSymbolDeclarationNode { Name: string name, Index: long index }:
                 return new SceneSymbol { Name = name, Index = index };
             case RecordSymbolDeclarationNode recordDeclaration:
                 return CreateRecordSymbolFromDeclaration(recordDeclaration);
